# Use Rc for arrays to improve performance