Every time an array is accessed, the entire array is cloned (even when we just want to read from the array):
https://github.com/BattleMage0231/rickroll/blob/4e7b7666355abb3cde36818a4017ed3d2aedac9f/src/util.rs#L99-L105

We can avoid cloning the array's `Vec<RickrollObject>` by wrapping it in an `Rc`. A benchmark using https://github.com/yaxollum/rickroll-quine/blob/main/Quine.rickroll shows an 8x performance improvement:

```bash
# on current master:
$ time rickroll Quine.rickroll > /dev/null

real	0m2,063s
user	0m2,046s
sys	0m0,013s

# after using Rc for arrays:
$ time rickroll Quine.rickroll > /dev/null

real	0m0,249s
user	0m0,241s
sys	0m0,008s
```